client1/requirements.txt
paho-mqtt
pyJWT[crypto]
pymongo[srv]
requests
python-dotenv
